+++ packages/sdk-react/src/components/full/video-conference/types.ts
@@ -17,6 +17,7 @@ export interface Features {
 	chat?: FeatureValue;
 	recording?: FeatureValue;
 	screenShare?: FeatureValue;
+	pictureInPicture?: FeatureValue;
 	whiteboard?: FeatureValue;
 	reactions?: FeatureValue;
 	handRaise?: FeatureValue;